Algorithm

StationarySheath.compute_ne! — Method

compute_ne!(ne, chare, tointe, params, f_eb, phix)

Computation of

\[ n_e(x) = \int_{v\in\mathbb{R}} f_e(x,v) dv = 2 \int_{v \leqslant 0} f_e(x,v) dv.\]

StationarySheath.compute_ni! — Method

compute_ni!(
    ni,
    chari,
    vi,
    vip,
    toint,
    params,
    f_eb,
    phix,
    phidx
)

Computation of

\[ n_i(x) = \int_{v\in\mathbb{R}} f_i(x,v) dv = 2 \int_{v \leqslant 0} f_i(x_b(x,v),v_b(x,v)) dv.\]

StationarySheath.integral_char_00! — Method

integral_char_00!(vi, toint, params, phix, f_eb)

Special case of the critical characteristic: use of a rectangle integration scheme to avoid the equilibrium point $(x=0,v=0)$.

StationarySheath.integral_char_vb0! — Method

integral_char_vb0!(
    vi,
    toint,
    params,
    ixb,
    phix,
    phidx,
    f_eb
)

Compute the integral of $f_e$ along the ion characteristic reaching $(x_b>0,v_b=0)$ at $\tau=0$. Here $x_b$ is known through $\phi(x_b)$ only. Use

\[ \int_{\tau=-\infty}^{0} f_e(x(\tau),v(\tau)) = \int_{v=vmin}^{0} \frac{f_e(x(v),v)}{-\phi'(x(v))} dv\]

and a trapeze integral on the points $(x(v),v)$ s.t. $x(v)\in$meshx. Preallocated vectors vi and toint will be modified by the function.

StationarySheath.integral_char_xb0! — Method

integral_char_xb0!(vi, toint, params, vb, phix, f_eb)

Compute the integral of $f_e$ along the ion characteristic reaching $(x_b=0,v_b<0)$ at $\tau=0$. Use

\[ \int_{\tau=-\infty}^{0} f_e(x(\tau),v(\tau)) = \int_{z=0}^{1} \frac{f_e(z,-\sqrt{v_b^2 - 2*\phi(z)})}{-\sqrt{v_b^2 - 2*\phi(z)}} dz\]

and a trapeze integral on the points $(z,-g_z(v_b))$ s.t. $z\in$meshx. Preallocated vectors vi and toint will be modified by the function.

StationarySheath.solve_Poisson! — Method

solve_Poisson!(phi, params, ni, ne)

Resolution by quadrature when $\phi$ is approximated pointwise.

StationarySheath.solve_Poisson! — Method

solve_Poisson!(phi, params, ni, ne)

Resolution by least squares when $\phi$ is decomposed in a polynomial basis.